`ix` considered harmful. Use `iloc` and `loc` instead
This completely tripped me up in class today when I did not use `index_col='country'` and I called `data.ix[0:10]`  
Because the row indexes were already numbers, `ix` was inclusive at both ends instead of the expected behavior of not including the final index.  This caused extreme confusion and because it was the last episode of the day, my students left confused.
Instead, we should use `iloc` everywhere we want to use numerical indices, and `loc` everywhere we want to use index names so that the indexing behavior is isolated to the function not the data type of the index.