switch to jekyll 3 instead of github pages
- otherwise jekyll serve generates errors along the lines of:

Liquid Exception: "<USERNAME>/<PROJECT>" is invalid as a repository
identifier. Use the repo/user (String) format, or the repository ID
(Integer), or a hash containing :repo and :user keys. in
/_layouts/base.html web_1  | jekyll 3.6.2 | Error:
"<USERNAME>/<PROJECT>" is invalid as a repository identifier. Use the
repo/user (String) format, or the repository ID (Integer), or a hash
containing :repo and :user keys.