Turning lists into narrative and adding context
Bulleted lists were turned into narratives. Additional context was added when necessary. 
Questions, objectives, and keypoints were modified to better reflect lesson purpose.
Titles modified to better address objectives.
The challenge "More Math" was moved below "Equations". This challenge feels like it does not belong with the others, because it introduces the concept of trying to run multiple codes in one cell, while the other challenges focus on Markdown. I would suggest either removing this challenge from this lesson, or adding a section that introduces this concept to participants more explicitly (either in its own section or as a "thinking ahead" section at the end).