Lunch reflection goals first draft
Added a first draft of three lunch time reflections modeled off of the morning coffee break reflections.
My intention was to get them thinking about what sort of packages they might find useful and that using a package can be much easier than writing all the code themselves. The second was to get them thinking about tidy data concepts and how they would have to have things to be machine readable. The third is to hint at things like loops and conditionals that are necessary to handle large and/or complex tasks.