Add a table for arithmetic operators
I was recently teaching this lesson and a student was confused about the arithmetic operators (younger student who likely had only, for example, used a calculator with the standard division sign). For that reason, I added a small table with the arithmetic operators before the first operation is performed, just for the students to have for reference.