Reorganisation of lesson
This is a proposal for restructuring the lesson so that more basic programming concepts such as conditionals, loops are presented earlier. These are basic concepts and should be earlier in the lesson. Later lessons now focus on functions, and then make use of data frames and plotting. Further updates to the material may be necessary to make this more consistent. But this flow is an improvement on the earlier flow.