Update 03-types-conversion.md
Feedback from a workshop recently was that using 'first' and 'second' for variable names was confusing "especially when you have to say things like “the first value of ‘first’ is … and the second value of ‘second’ is …”. This change is to suggest changing those variable names to 'variable_one' and 'variable_two'. Maybe alternatively x and y, or v1 and v2 - basically anything other than first and second! 

Similarly, I suggest renaming the variable 'complex' - again this caused confusion with the variable name being the same as the type name.

Thanks :-)